Create migration

+++ TingStoreAPI/Models/Order.cs
@@ -20,7 +20,7 @@ namespace TingStoreAPI.Models
         public DateTime orderDate { get; set; }
         [Column(TypeName = "Money")]
 
-        public long TotalAmount { get; set; }
+        public decimal TotalAmount { get; set; }
 
         public int orderStatusId { get; set; }
